Rename `TableInfo` to `TableDTO`
